+++ meetup2apricot/event_registration_type.py
@@ -1,5 +1,7 @@
 """Wild Apricot event registration types for free and paid registrations."""
 
+from .exceptions import NoRestrictionsMatchError
+
 
 class EventRegistrationType:
 
@@ -15,6 +17,7 @@ class EventRegistrationType:
         is_enabled=True,
         description="",
         availability="Everyone",
+        guest_registration_policy="Disabled",
         **more_attributes,
     ):
         """Initialize with parameters for the Wild Apricot API."""
@@ -27,6 +30,7 @@ class EventRegistrationType:
         self.description = description
         self.availability = availability
         self.more_attributes = more_attributes
+        self.guest_registration_policy = guest_registration_policy
 
     @property
     def display_count(self):
@@ -48,7 +52,7 @@ class EventRegistrationType:
             "GuestPrice": self.price,
             "Availability": self.availability,
             "MaximumRegistrantsCount": self.maximum_registrants_count,
-            "GuestRegistrationPolicy": "Disabled",
+            "GuestRegistrationPolicy": self.guest_registration_policy,
             "UnavailabilityPolicy": "ShowDisabled",
             "CancellationBehaviour": "AllowUpToPeriodBeforeEvent",
             "CancellationDaysBeforeEvent": 2,
@@ -81,25 +85,26 @@ class EventRegistrationTypeMaker:
     ):
         """Make a Wild Apricot event regirstation type appropriate for an event
         title."""
-        restriction = self.choose_event_restriction(event_title)
-        if restriction:
+        restriction = self.choose_event_restriction(event_title, price)
+        if restriction.member_levels:
             return self.make_restricted_apricot_type(
                 event_id, maximum_registrants_count, price, restriction
             )
         else:
             return self.make_unrestricted_apricot_type(
-                event_id, maximum_registrants_count, price
+                event_id, maximum_registrants_count, price, restriction
             )
 
     def make_unrestricted_apricot_type(
-        self, event_id, maximum_registrants_count, price
+        self, event_id, maximum_registrants_count, price, restriction
     ):
         """Make an unrestricted event registration type for Wild Apricot RSVPs."""
         return EventRegistrationType(
             event_id=event_id,
-            name="RSVP",
+            name=restriction.name,
             price=price,
             maximum_registrants_count=maximum_registrants_count,
+            guest_registration_policy=restriction.guest_policy,
         )
 
     def make_restricted_apricot_type(
@@ -117,15 +122,20 @@ class EventRegistrationTypeMaker:
             availability="MembersOnly",
             maximum_registrants_count=maximum_registrants_count,
             AvailableForMembershipLevels=member_levels_for_json,
+            guest_registration_policy=restriction.guest_policy,
         )
 
-    def choose_event_restriction(self, event_title):
+    def choose_event_restriction(self, event_title, price):
         """Choose and return the first event restriction with a pattern that
-        matches the event title. Return None if no patterns match."""
+        matches the event title and price category.  Note that the event
+        restriction list ends with a default that should match all events."""
         for restriction in self.event_restrictions:
-            if restriction.pattern.search(event_title):
+            if restriction.pattern.search(event_title) and restriction.matches_price(
+                price
+            ):
                 return restriction
-        return None
+        message = f"Bug: No restrictions match {event_title=!r} {price=}"
+        raise NoRestrictionsMatchError(message)
 
 
 # vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4 autoindent