+++ HISTORY.rst
@@ -15,6 +15,12 @@ Recent History
 .. Next Release
 .. ------------------
 
+1.9.0 (2021-08-10)
+------------------
+
+* Add registration restrictions based on price.
+* Make guest registration policy configurable.
+
 1.8.1 (2021-04-09)
 ------------------